# api/openapi.yaml
openapi: 3.0.3
info:
  title: ErgPower API
  version: 0.1.0
  description: >
    Live rowing data from a Concept2 PM5, for a browser dashboard. v1 is read-only: connection status,
    a full current-state snapshot, and a multiplexed Server-Sent Events stream. All physical quantities
    are SI (metres, seconds, watts, Newtons, bpm); pace is seconds per 500 m. Commands
    (connect/disconnect, device scan) and stored-session browsing are deferred to later revisions.
servers:
  - url: /api/v1
...
paths:
...
  /coach/progress:
    post:
      operationId: coachProgress
      summary: Progress coaching over an explicitly selected set of sessions (only when configured).
      description: >
        Grounded progress coaching across the chosen sessions — the most recent is the current piece and
        the rest form the baseline. Optional — 409 when no provider is configured.
      tags: [coach]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProgressRequest" }
      responses:
        "200":
          description: Progress coaching text and the model that produced it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoachResult" }
        "404":
          description: A selected session was not found.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: No LLM provider configured.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
  /sessions/{id}/coach:
    get:
      operationId: getSessionCoach
      summary: LLM coaching for a stored session (only when a provider is configured).
      description: >
        Grounded natural-language technique coaching built from the session's deterministic analysis
        (features + scorecard + flags) and a Kleshnev rubric. Optional — 409 when no provider is
        configured. The model never sees raw curves.
      tags: [coach]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: mode
          in: query
          required: false
          description: >
            single (default) coaches this session alone; progress also grounds the athlete's recent
            same-type history from the cross-session index and narrates improvement over time (falling
            back to single when there is no comparable history).
          schema: { type: string, enum: [single, progress], default: single }
      responses:
        "200":
          description: Coaching text and the model that produced it.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoachResult" }
        "404":
          description: No such session.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "409":
          description: No LLM provider configured.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
...
components:
  schemas:
...
    CoachResult:
      type: object
      properties:
        model: { type: string, description: "The model that produced the coaching." }
        text: { type: string, description: "Natural-language technique coaching." }
      required: [model, text]
...
    ProgressRequest:
      type: object
      description: A selected set of sessions to coach progress over (most recent = current piece).
      properties:
        sessions:
          type: array
          items: { type: string }
      required: [sessions]
...
    Problem:
      type: object
      description: RFC 7807 problem details.
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
      required: [title, status]